Internal vs external routing fixes on sign in form. set page height properly when popping popups. Never set the container hash when in embedded comments.
Always make the signin form use the inner url. Javascript takes care of
setting the appropriate target & top when registering. Set the iframe
height immediately when popups are loading/revealed. Fixed a bug that
caused parent location to be set in the hash even in embedded comments
(it should never happen there).