Hash root user pass upon installation.
There's an oversight with Vanilla's installer where upon initial user creation, the password is stored in plaintext in the DB. This can only be corrected via changing the password. This changeset corrects this oversight as it can be easily overlooked.

Special thanks to Phillip Stephens (https://wiiking2.com/) for noting this behavior.