Update FilterForm of class.model.php
Add 'Format' to the fields that are filtered out from POST data.   
   
By now it is possible to spoof the post data to set Format to whatever you like. So malicious users can override Garden.InputFormatter settings on individual basis. Although I think that this is no big issues, users shouldn't be able to add arbitrary text to columns where they shouldn't be able to insert *anything* on their own