Proposal: Allow simple javascript anchors
This change would allow Anchors to be simple javascript links, often useful when using an anchor simply to trigger a javascript event.
Currently attempting to create an anchor such as:
<a href="javascript:;">I can't exist</a>
Results in Garden prepending the forum path to the link, e.g. 
<a href="/vanilla-forums/javascript:;">I can't exist</a>

This may just be a personal preference of mine, but it seemed like useful functionality to have without any sacrifice in particular. Figured it was at least worth the suggestion.