Fix jquery atwho implementation
- Options in the @mention flyout can now be selected with a mouseclick.
This was already the librarys default, but "overridden" by the
display_timeout setting of 0 in global.js. (The inputs blur event is
triggered before the click event, closing the flyout before, anything
can be selected. This is the reason for the timeout being there.)

- The tab key no longer closes the flyout, but moves to the next item.
This required 2 minimal changes in the library.