Change the way auto-increment fields are validated.
Previously auto-increment fields would never be validated. Now because of sparse insert support they can be validated, but treated as if they have a default value which they do. This allows auto-increment fields to be type checked on updates and ensures that their values show up in `Gdn_Validation->ValidationFields()`.