Don’t show misleading permissions when editing a role without any permissions.
If there are no entries for a role in `GDN_Permission` then editing that role shows default permissions. This is misleading because that role really has no permissions. This fix makes the role show no permission so that admins know to fix the role.

This is an issue that usually won’t occur unless the data is incorrect from something like an import.