Make title(), description(), error- and informMessage() translatable by default.
Add the t() function inside the controller methods so that the methods texts are translatable by default. Changing this will lead to some double translations throughout the code. I do not think that this will effect performance too badly, since the methods are used sparsely and not in loops or anything like that.

Also fixed casening (completely?) in class.controller.php - even in comments!