AllIPAddresses field size 100 -> 1919
At the time when IPv4 was the only supported standard, which, when storing the human-readable format, is a maximum of 15 characters long, I believe that the max size of this field (100) was arbitrarily chosen.

100 / 15 = 6.66 repeating. Not very intuitive.

Given that this update will be supporting IPv6 addresses which have a maximum character range of 39 when storing the human-readable format, I've tried to be a bit more thoughtful with my choice on this field size.

For instance, given how ISPs are or will be handling the assignment of IPv6 addresses in the future, it's reasonable to expect that a single person would have access to ~48 of these addresses throughout their tenure. Give or take.

Doing a bit of simple math: 39 * c + (c-1)
Where c is the estimated maximum count of addresses. The bit at the end accounting for the separator.

For planning to store 48 addresses, this would be: 39 * 48 + (48-1) OR 1919