Update class.discussionmodel.php
This will allow plugin owners more control over the sorting order, specifically to have multiple sort orders.  Typical example: sort so that bookmarked discussions are on top of other discussions, alerted discussions (see DiscussionAlert plugin) ahead of bookmarked and ahead or the rest which should be in the "standard" order.