Minor coding improvements to class.form.php
In method date() array for days and years were initialized as empty arrays and right the next step was to add their first element with a static value. I've changed that so that the are initilized with that static element right from the start.

I've also changed all
`public function ... (..., $Attributes/$Options = false)` to
`public function ... (..., $Attributes/$Options = array())`.
The parameter is expected to be an array, so its default should be an array, too.