Make proxyRequest::curlHeader handle responses with multiple identical fields correct
When requesting the HEAD of a site that has e.g. multiple "Link:" lines, only the last of those lines made it to proxyRequest->ResponseHeaders array. It is a flat array so `proxyRequest->ResponseHeaders['Link'] = ...` is called for every Link line.

I've made it do a check for double entries and to transform the entry to an array if needed.