Implement an alter table threshold
Add an optional alter table row threshold which represents the maximum number of rows a table is allowed to have for an in-application alter to be allowed to run. If a table is too large then the admin is expected to copy the SQL from the structure endpoint into their RDBMS.

By default this threshold is turned off, but may be turned on by config.