Add the “Dir” and “IconUrl” keys to plugin info and use them
The “Dir” key defines the application root relative directory of the plugin. This is useful to build both file paths and URLs for the plugin’s assets. The “IconUrl” key defines the application root relative path of the plugin’s icon, if any, and can be passed right to asset().

These two keys will be standardized across all addons.