Correct field length of GDN_Tag.FullName
**Please be careful with this!**

There has been an error that made the field FullName hold 255 characters right after installation until the first run of /utility/structure. Then it is cut down to 100 characters.
I don't know if the model makes any restrictions, but if not, the FullName could be user for strings longer 100 characters and thus might get cut when utility/structure is run for the first time.

It might be safer to set all values to 255 based on that error.