Make "AS" in "select X as Y" case insensitive
The regex which searches for the SQL keyword "AS" in the SELECT string is not case insensitive. `->select('Name AS UserName`')` would output a column named "AS UserName".

I've also stripped down the rest of the if condition, where there was an unneeded regex and I've tested it with following examples:
~~~
$test[1] = 'ColumnWithoutAlias';
$test[2] = 'Column   `AliasWithoutAs`';
$test[3] = 'Column    as AliasLowerCaseAs';
$test[4] = 'Column AS AliasLowerCaseAs';
$test[5] = 'Column aS     AliasMixedCaseAs';
$test[6] = 'Table.`Column`';
$test[7] = 'Table.Column';
~~~


The regex used for the from() method is case insensitive, just in case you thought about checking that ;-)

They all show the expected result.