Refactor the dispatcher
This is a fairly big refactor, but is really only moving things out of large methods into smaller methods.

There is a bit of a methodology change too where controllers are discovered before their methods are examined. This mirrors where the dispatcher is going even though it currently doesn’t serve much of a great purpose.