Fix asset sorting
The dispatcher refactor assigns a lowercased application name to the controller rather than the title cased one. This causes the asset sort arrays to not work. Rather that changing the application that’s assigned to the dispatcher I opted to make the asset sort containers case-insensitive.

This PR also includes some variable renaming to be in line with code standards.