Add BeforeCheckPreference event to ActivityModel
While trying to write a plugin which uses Pushbullet for sending notifications to users, I've found no way to queue activities for users who have neither the Email nor the Notification preference set, based on the rigid `if (!$Popup && !$Email) return` preference check
I needed an event before that check so that I could queue an activity based on my custom preferences ("BeforeCheckPreference"). I would have prefered a "BeforeQueue" event right before the `self::$Queue[...] = ...` but that would have required a logic change in that method and I wanted to be least invasive.

What would it take to ensure this change is in your next release (even if it would be a 2.2.bugfix)?