Enhance isUrl() to also identify data urls as urls
I've stumbled upon this when I was trying to write a plugin that generates a SVG avatar. Using userPhotoDefaultUrl() to return a data url does not work, since isUrl() doesn't recognize this format and prepends the upload path before the data url.

The isUrl() function can only be overwritten in the bootstrap.before.php which is quite uncomfortable for plugins.

I'm not sure if this would cause any problems anywhere, but as far as I know Vanilla doesn't use inline images.

But in the profile the isUrl() is sometimes used to determine if the profile picture is a "$RemotePhoto". This would give a false positive, though...