Fix profile/discussions more pager
On the profile/discussions page, the more pager appears in the container where the paging occurs. The morepager configuration currently appends additional discussions to the end of the container. This results in the more link being anchored in the middle of the discussion list rather than at the end. 

This PR configures morepager to prepend the discussions to the more link. It also removes the unneeded not(.Message .MorePager) filter.