Give themes the option to set their layout
This allows a theme to set its expected layout in its about.php array. This layout will be saved to config when the theme is enabled. 

When previewing a theme, it will preview with the expected layout.

To test, edit the about.php of a theme and add
```
'Layout'      => [
    'Categories'  => 'modern',
    'Discussions' => 'modern'
]
```
Clear your cache and then enable the theme. Ensure that the layout on `/settings/homepage` has changed to what has been specified in the theme. Preview the theme and make sure that no matter what the config is set as, the theme previews with the settings specified in about.php.

Closes https://github.com/vanilla/vanilla/issues/4666