Remove the ill-named js-hijack class from confirm modals
The js-hijack class was used to signify that a modal's endpoint should be processed using ajax and to add the transient key to the request. However, most every time we use a confirm modal (js-modal-confirm), we also want the js-hijack functionality. This PR removes the js-hijack class. It adds a data-follow-link attribute that simply follows the link provided rather than processing the endpoint for the rare circumstances we want that to happen.

The js-hijack class also is apt to be confused with the existing Hijack class in Vanilla.

Partially addresses https://github.com/vanilla/vanilla/issues/3937