Add the Cookie class
The Vanilla\Web\Cookie class is a utility class for handling reading
and changing cookies throughout the lifecycle of the request.

Currently, we don’t have controlled cookie management. Just a bunch of
ad-hoc calls to **$_COOKIE** and **setcookie()**. This results in
problems such as unnecessary Set-Cookie headers.

What still needs to be done:

- [ ] Add unit tests.
- [ ] This implementation may have bugs, the above unit tests should
uncover them.

When implemented this class will be shared in the container and then
any class that wants to manipulate cookies can depend on it.