Allow the database to be instantiated when there is no default config.
This is a bug that was uncovered during unit testing since Vanilla will always have a default config. Basically the following call:

```php
$defaultConfig = Gdn::config(‘Database’);
```

This returns **null** when there is no default config and then makes all the following array merges to come out null too.

Fixing that one line fixes the bug, but I couldn’t let this go without cleaning up the code. On inspection I think this is the first function I touched on Vanilla and represents a very novice understanding of the PHP language. It’s quite the nostalgia trip to look at it, but we certainly don’t need so many calls to val() anymore.