Move Private Community plugin to core
The Private Community plugin existed only to manage whether the "Enable Private Communities" toggle is shown on the Roles page in the dashboard, which manages the Garden.PrivateCommunity config setting. Enabling the plugin doesn't automatically enable the setting, so it can be a little confusing. 

This makes it unnecessary to enable the plugin to manage the config and automatically adds the toggle to the Roles settings page. It also removes the plugin.

Users that already have the plugin enabled will not see a functional change. Users that do not have the plugin enabled will see the "Enable Private Communities" toggle on their Roles settings page.