Fix issue that users could see and post to conversations they have already left
When users know the url to a conversation they have left, they can still see it by simply visiting the url. They could also post to such a conversation which "re-injects" them into such a conversation.
The fix in the ConversationModel disallow users to see left conversations and the fix in the ConversationsModel disallows them to post to such a conversation (with a spoofed request).