Increment duplicate form element IDs when multiple are added
It is invalid HTML to have duplicate IDs. Although we should consider
getting rid of default element IDs altogether that would be too much of
a change so this change at least increments the IDs if they are re-used.

A few points:

* The first use of an ID should not add any prefix. This means that
unique IDs will behave the same way they did before.
* If you view the source on a page that has duplicate IDs you’ll notice
that some of the IDs seem out of sync. Look at the Form_hpt in
particular. I decided that this is acceptable since we aren’t looking
for an aesthetic solution to a dirty problem. Again, the ultimate
solution should be to remove form element IDs altogether.
- There is a lot of javascript targeting form IDs. This seems risky and
some testing should be done. However, according to the [jQuery docs on
the ID selector](http://api.jquery.com/id-selector/) only the first
element with a given ID is queried no matter what so the behavior
should not change.