Replace quotes in the promoted content module.
The Promoted content module was showing quotes in the text, and since it only provides a small excerpt, having a quote at the beginning of a post would present the quote as the actual content. I've stripped them the same way that we strip out spoilers.

I think this should probably be the default behaviour. Since plaintext format doesn't carry any of the actual information about the quote it should be replaced all the time. Off the top of my head it doesn't seem like that would cause any issues, especially since we already do replacing of spoilers in plaintext.

If you think it would be a good idea to always replace the quotes I can remove the extra parameter and the if statement.