Don’t render the numbered pager when the current page is off the end of the pages
This fixes a bug where the numbered pager would render in a bizarre manor when the current given page is greater than the total number of pages. Now it just renders a single previous link that puts the user back to the last page.

Remove last page reset code

This is odd code that renders the last page twice under certain circumstances.

Render a not found page if the page number is too high

Don’t add head links if the page number is invalid

Throw 404 if past the end of the pages on category listings

Throw 404 if past the end of the pages on recent discussions

Revert "Remove last page reset code"

This reverts commit 1b7b1f0d309ff6b0538975e61b910d9d0c78fb16.

Fix not found error when there are no comments

Fix the offset only when invoking the auto offset

Don’t render the single pager link when the pager wouldn’t normally render at all

Don’t reset the offset when auto offset is not being invoked

The first page should always be found

Refactor the page checking logic into method