Always load the transient key
The session was only loading the transient key when there is a valid user even though the TK cookie was submitted.

We need the TK for login pages where there is no session so the TK is important. As it turns out the entry controller always ensured the TK itself, but this really should be handled in the session.

The fix in Gdn_Session::start() caused a nasty redirect loop when signing a user out because Gdn_Session::loadTransientKey() had a bug. This shouldn’t be an issue now, but shows that this seemingly simple fix can cause some bad regressions because of the state of Gdn_Session.