Relax CSRF checks for AJAX requests
Add the ability for non-CORS AJAX requests to validate if they pass the X-Requested-With header.

A request is now authenticated if it passes an X-Requested-With header and the Origin and Referer headers match the host. There is an exception for a missing Origin header to handle a glitch in IE11.

This is based on the OWASP article on [CSRF prevention](https://www.owasp.org/index.php/Cross-Site_Request_Forgery_(CSRF)_Prevention_Cheat_Sheet).

Since this technique only works for AJAX requests the transient key verification is being left in.