Look for ArgumentCountError exceptions when calling event handlers in Gdn_Pluggable
This is a temporary change to give us some protection when moving to PHP 7.1 because it has promoted this warning to an error. We think we don’t really have handlers that express the bug, but if we do we don’t want them to break pages. With this change they will log errors and we can fix them ASAP.

Once we’ve caught errors this change should be reverted.