Add default parameters to Gdn_ErrorException and fix cleanErrorArguments()
With PHP 7.1 functions can no longer be called with too few arguments. This was causing issues with the Gdn_ErrorException constructor. I copied the defaults of its base class to help with this.

While fixing this issue I uncovered an issue with the cleanErrorArguments() function. It has no protection for infinite recursion and also doesn’t clean the security blacklist on object properties. I fixed this, but decided to make another version because I wanted a no side-effects version of the function.