Remove vine’s external script
TLDR: Removing vine’s script seems to still allow their embeds to work minus autoplay. Since vine is defunct this is an acceptable tradeoff.

Including an external script for vine means that it can execute whatever code it wants on the page. This poses a security concern that we usually overlook for major sites.

However, another issue is tracking cookies. Sites are not able to block cookies set in such scripts. Making vine just an iframe allows scripts to that want to block all iframes also apply to vine. Since vine is defunct and only in archive mode it shouldn’t really be a holdout for scripts.