Add the CacheControlMiddleware and tweak default caching
This PR is an attempt to put the application’s cache headers inline with our Varnish overrides so that the application has more control over browser and proxy caching.

- Signed in users don’t cache by default. This behavior is unchanged.
- Guests cache requests for two minutes.
- The CacheControlMiddleware handles caching for the API.
- Since endpoints can now opt-out of caching the /authenticate, /utility and /entry endpoints now do so.
- HTTP 1.0 headers are removed and added when headers are sent. This allows programmers to only code with just the Cache-Control header.

This is a fairly significant change so I’m requesting people from Ops also review the PR. We should also test this on infrastructure before merging.