cleanup: fix license path and content
This patch moves the license file to LICENSE and provides the raw
text license file which will result in GitHub being able to cleanly
parse the license for the project, which will also fix the badge
which shows 'unknown' license presently.

It also fixes the URL which was pointing at a deleted branch and
instead points at the LICENSE in master.

This change make no changes to the licensing, but only to the style
of the file and path.