A config overide for relative/absolute path in setCacheDir
I used `Cache.` config prefix becuase it already exists. I didn't use `Cache.Filecache.` becuase this is a lower level thing than filecache, and involves the object/class/addon caching regardless of type of cache.

Default is relative for backwards compatibility. 

Setting false will allow you to set an absolute path to the cache directory, as a constant in a similar way to the upload directory.  Useful for multi-tenancy arrangements.