Add state token support to Gdn_OAuth2
This PR uses the verification tokens from `SsoUtils` in `Gdn_OAuth2` to prevent CSRF and replay attacks when signing in via OAuth 2. Notes:

- `SsoUtils` is injected via the `Vanilla\InjectableInterface` to avoid affecting subclasses.
- The sign in has been changed into a redirect to avoid HTTP cache pollution. Even though we don’t fully respect cache headers, I’ve added the appropriate no cache headers before the redirect too.
- The state has been changed from a URL encoded array to a base 64 JSON array because I found that some OAuth servers can’t handle the double encoding.

Testing

I tested this with two subclasses of `Gdn_OAuth2`: The Google Sign-in addon and a custom subclass. Both worked and also threw the appropriate error when the token was wrong.

Closes https://github.com/vanilla/vanilla-patches/issues/564