Make redirectTo() also work with AJAX requests
We have a long standing difficulty when converting regular pages to AJAX pages where everything works except for redirects. Currently we go through and change calls to `redirectTo()` to `Gdn_Controller::redirectTo()` and have had to do that over and over again.

This change proposes putting some logic within `redirectTo()` itself to properly render the correct redirect code if the call is being made within an AJAX request. This is a bit of a kludge, but I figured I would put this out for code review because it is a trade off between a kludge and constantly doing the same bug fix over and over again.