Escape the title in Gdn_Theme::logo()
When a site doesn’t have an uploaded logo image then the site’s title is used. However, the title is not escaped. Even though the title is permission gated it is a good practice to escape the title in this function nonetheless.

I polled all of our sites and see that the only sites that would be affected all have “&” in their titles.

Closes https://github.com/vanilla/vanilla-patches/issues/391.