Add support for private communities in APIv2
Previously, private community functionality was checked before the APIv2 was invoked. This change adds separate private community processing on the APIv2 with permission bans.

Now, when in a private community without a session the user is banned, but individual endpoints can opt-out of the ban which is useful for authentication endpoints.