Allow re-authentication if a user has a Vanilla password, even if the registration is set to connect
This will allow the narrow case where a user might have a Vanilla password, but then have connect only registration. This prevents a lockout for such users.

There is still the case where users may have a different password from an import, but that will exhibit the previous behavior. If customers report such a problem then we can expand the check to all non random passwords.