Refactor embed filters to be pluggable
I've refactored the the embed filtering being done in `Quill\Filterer` to be handled be pluggable filters on the EmbedService.

The filter method was starting to get quite long, and since filtering logic tends to be embed specific, I want embeds to be able to plug in and add their own filtering.

This has an added benfit of making filtering a specific type of embed data capable of happening in the embed itself.