Make the controller throw a ServerException rather than a generic Exception
Currently, the `Vanilla\Web\Controller` throws a generic `Exception` when the session hasn’t been injected. If you are using PHPStorm then this riddles subclasses with a tonne of code inspection warnings which makes code hygiene quite difficult.

This change proposed to change the exception type to our general HTTP level `ServerException`. It’s the same code, but can then be ignored in your project settings. This makes sense, because we generally expect `HttpException` subclasses to throw their exceptions up to the dispatcher.