Add a recursion guard in Gdn_Format::to()
We have an ill-advised use of `Gdn_Format::to()` where we allow arrays or objects to be passed and the class will format everything in them. Although the practice is heavily discouraged, we still support it.

There is a problem with self-referencing objects or arrays where `Gdn_Format::to()` will infinitely recurse when formatting in that case. This fix puts a guard against infinite recursion by keeping track of the objects and arrays that have already been formatted and not re-formatting them.

This also adds core support for the “display” format as a sanitized format. That’s our version of `htmlspecialchars()` with some extra processing. I think that’s an okay thing to do, but am asking reviewers to check it out.