Fix some lint errors and code coverage
Now that our code coverage of functions.general.php is in good shape I’ve added some ignore directives for areas that we just can’t hit.

In doing that I wanted to clean up some lint errors because I find it easier to not introduce new ones if an entire file is in a clean state.

Our old `Gdn_Autoloader` class has not been used for some time so I’m marking it as deprecated.