Remove the fuzzy category watch from the discussion model
The discussion model has some logic to read a select set of discussions in order to try and mark the category read. I’m not super sure this has ever really worked, but is troubling from a scalability point of view.

This PR puts that logic behind a feature flag so that we can try a deploy with it turned off, but flip it back on if there really is some real functionality required.